+++ src/main/java/com/brewmes/demo/model/Machine.java
@@ -317,7 +317,7 @@ public class Machine {
      * @param beerType  the beer type for the machine.
      * @param batchSize the amount of beer to produce until stopping
      */
-    public void setVariables(int speed, BeerType beerType, int batchSize) {
+    public void setVariables(double speed, BeerType beerType, int batchSize) {
         try {
             //Set beertype on the machine
             NodeId setBeerType = new NodeId(6, "::Program:Cube.Command.Parameter[1].Value");
